requirements.txt
Django==5.1.4
python-telegram-bot==21.10
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
